migrate all existing testing over to jest
