--- libs/libskynet/tsconfig.test.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2021",
-    "declaration": true,
-    "module": "esnext",
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "outDir": "./dist-test"
-  },
-  "include": ["src-test/**/*"]
-}
